Enable the fields in the Profile Panel to wrap if it is too long
The Profile Panel is cutting off very long but valid names and address
and there is no other way to view them, even after resizing the window.

Let's enable the set the max width of the VBox within the Profile Panel
and enable to fields within to wrap around.

Setting the max width smaller than the min width of the Profile Panel
and enabling wrapping for the labels will ensure that the fields will
not be cut off and be visible at least in some window size.